[CODE] Snake resource changes
Changed the position of the snake so that it is in the main folder.  Also added some config variables so that it can accept different file endings. Added a if statement to check if RepeatMasker files are there and makes changes in last rule to do the bedtools intersects.   Added paths to files (python files and config file) so that it can be run locally using -v/LSF_DOCKER_EXPORT